Do not try to compile if the QtWebEngine is missing
The Qt Web Browser is now part of Automotive super repository
(tqtc-boot2qt/b2qt-super) and also part of the Automotive releases.

We don't want to make the qtwebbrowser to fail b2qt-super integration
in Coin/CI if some platform configuration does not support qtwebengine.
We just want it to ignore that specific build.

Change-Id: Ifc0b85e2e8617f7c6bbd6e6a1a5d6282cb8de75f
Reviewed-by: Kai Koehne <kai.koehne@qt.io>